+++ src/app.module.ts
@@ -4,7 +4,7 @@ import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
-import { LoginController } from './login.controller';
+import { LoginController } from './controllers/login.controller';
 import { LoginService } from './login.service';
 
 @Module({
